rename unifi.py to coordinator.py
In keeping with the [developer docs](https://developers.home-assistant.io/docs/creating_integration_file_structure), the unifi.py file which contains subclasses of DataCoordinator and CoordinatorEntity should be renamed to coordinator.py
